+++ reikna/matrixmul.py
@@ -109,7 +109,8 @@ class MatrixMul(Computation):
                     global_size=(grid_width * block_width,
                         blocks_per_matrix * basis.batch * block_width),
                     local_size=(block_width, block_width),
-                    render_kwds=render_kwds)
+                    render_kwds=render_kwds,
+                    dependencies=[('out', 'a'), ('out', 'b')])
             except OutOfResourcesError:
                 continue